[CHANGE] Move all template tags into one file
