+++ afat/templates/afat/bundles/afat-css.html
@@ -1,4 +1,4 @@
-{% load afat_versioned_static %}
+{% load afat %}
 
 <link rel="stylesheet" href="{% afat_static 'afat/css/aa-bootstrap-fix.min.css' %}">
 <link rel="stylesheet" href="{% afat_static 'afat/css/afat.min.css' %}">
