+++ afat/templates/afat/bundles/afat-fatlink-details-js.html
@@ -1,3 +1,3 @@
-{% load afat_versioned_static %}
+{% load afat %}
 
 <script src="{% afat_static 'afat/javascript/afat-fatlink-details.min.js' %}"></script>
