+++ afat/templates/afat/bundles/chartjs-js.html
@@ -1,3 +1,3 @@
-{% load afat_versioned_static %}
+{% load afat %}
 
 <script src="{% afat_static 'afat/libs/Chart.js/2.7.2/chart.min.js' %}"></script>
