+++ afat/templates/afat/partials/month_navigation.html
@@ -1,4 +1,4 @@
-{% load filters %}
+{% load afat %}
 {% load i18n %}
 
 <div class="afat-select-time-period afat-select-month">
