+++ afat/templates/afat/partials/statistics/overview/tabs/mine.html
@@ -1,4 +1,4 @@
-{% load filters %}
+{% load afat %}
 {% load evelinks %}
 {% load i18n %}
