+++ afat/templates/afat/view/statistics/statistics_corporation_year_overview.html
@@ -1,7 +1,7 @@
 {% extends "afat/base.html" %}
 
 {% load i18n %}
-{% load filters %}
+{% load afat %}
 
 {% block page_title %}
     {% translate "Fleet activity tracking statistics" as page_title %}
