+++ afat/templates/afat/view/statistics/statistics_overview.html
@@ -2,7 +2,7 @@
 
 {% load i18n %}
 {% load evelinks %}
-{% load filters %}
+{% load afat %}
 
 {% block page_title %}
     {% translate "Fleet activity tracking" as page_title %}
